requirements.txt
python-dotenv==0.20.0
aiogram==2.21
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.26.0
requests~=2.28.1
bs4==0.0.1
lxml~=4.9.1
beautifulsoup4~=4.11.1
APScheduler~=3.9.1
tzlocal~=4.2
typing~=3.7.4.3
loguru~=0.6.0